Moved styled component to their own file
